# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DVinyl API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Mobile/native client API for a DVinyl instance, covering authentication,
    collection, wishlist and sharing management, the full item lifecycle, plugin
    metadata, instance administration, and account settings.

    All routes are mounted under `/api/v1`. Requests and responses are JSON
    (`application/json`), except the file-upload endpoints (`multipart/form-data`),
    the share-link QR endpoint (`image/png`) and the import/export endpoints, whose
    exports answer CSV, ZIP or JSON file downloads.

    Authentication is a JWT Bearer access token obtained from `POST /auth/login`.
    Access tokens live 15 minutes (`expiresIn: 900`); use the rotating refresh
    token from `POST /auth/refresh`. Collection-scoped routes are additionally
    gated on the caller's role in the collection named in the path
    (`viewer` < `editor` < `admin`); instance admins pass every collection role.
    Admin routes require `isAdmin` on the account.

    Error envelope: every failed response is `{ "success": false, "error": "..." }`
    with the HTTP status appropriate to the cause. Success responses are
    resource-shaped (e.g. `{ "item": ... }`, `{ "items": [...], "page": ... }`).

    On an instance that has not completed first-run setup, every `/api/v1`
    request returns `503` with `{ "success": false, "error": "Instance not set up yet" }`.
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /collections/{id}/value-history:
    get:
      tags: [Collections]
      operationId: getCollectionValueHistory
      summary: Get a collection's value-estimate history
      description: |
        Requires at least `viewer` role on the collection. Returns the recorded
        value snapshots (`PriceHistory`) for the caller's own currency, oldest
        first. Values are never converted between currencies; `otherCurrencies`
        lists currencies the collection also has history in, so a client can
        explain an empty chart.
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      responses:
        '200':
          description: The caller's value history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValueHistory'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /collections/{id}/value-estimate:
    post:
      tags: [Collections]
      operationId: startCollectionValueEstimate
      summary: Start a collection value estimate
      description: |
        Requires at least `editor` role (the run records a history snapshot). Starts
        a background Discogs pricing run over the collection's enabled,
        estimate-capable items and returns `202` immediately. Poll
        `GET /collections/{id}/value-estimate/{jobId}` for progress and the result.
        Only one run may be active per collection; a second start returns `409` with
        the active job.
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      responses:
        '202':
          description: The run started.
          content:
            application/json:
              schema:
                type: object
                required: [estimate]
                properties:
                  estimate:
                    $ref: '#/components/schemas/ValueEstimateJob'
        '400':
          description: No enabled estimate-capable items in the collection.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { success: false, error: 'No estimable items' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: A run is already active for this collection.
          content:
            application/json:
              schema:
                type: object
                required: [success, error, estimate]
                properties:
                  success: { type: boolean, enum: [false] }
                  error: { type: string }
                  estimate:
                    $ref: '#/components/schemas/ValueEstimateJob'
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /collections/{id}/value-estimate/{jobId}:
    get:
      tags: [Collections]
      operationId: getCollectionValueEstimate
      summary: Poll a collection value estimate
      description: |
        Requires at least `viewer` role on the collection and that the caller
        started the run (anyone else gets `404`). Returns the run's status and
        progress; once `done`, `result` carries the totals and whether a history
        snapshot was saved. An unknown, expired, other-collection or other-user job
        is `404`.
      parameters:
        - $ref: '#/components/parameters/CollectionId'
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Current run state.
          content:
            application/json:
              schema:
                type: object
                required: [estimate]
                properties:
                  estimate:
                    $ref: '#/components/schemas/ValueEstimateJob'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Unknown, expired, other-collection or other-user job.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { success: false, error: 'Estimate not found' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /items/{itemId}/estimate:
    get:
      tags: [Items]
      operationId: estimateItemPrice
      summary: Estimate one item's price
      description: |
        Requires at least `viewer` role on the item's collection. Returns a single
        Discogs price for the item's external id. `condition` is optional and falls
        back server-side to VG+ when unrecognised. `estimate` is `null` with
        `reason: unavailable` when the provider has no price; a provider failure is
        `502`.
      parameters:
        - $ref: '#/components/parameters/ItemId'
        - name: condition
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: A price, or `null` when none is available.
          content:
            application/json:
              schema:
                type: object
                properties:
                  estimate:
                    oneOf:
                      - $ref: '#/components/schemas/PriceEstimate'
                      - type: 'null'
                  reason: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Unknown/non-visible item, or its type has no price provider.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '502':
          description: The price provider failed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /collections/{id}/settings:
    get:
      tags: [Collections]
      operationId: getCollectionSettings
      summary: Get a collection's settings
      description: |
        Requires `admin` role on the collection. Returns modules, visibility,
        personalisation and plugin settings as one JSON resource. `theme` is a
        `collectionType`/`home` -> preset-key map, and `modules`/`theme` always cover
        every registered plugin.
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      responses:
        '200':
          description: The collection's settings.
          content:
            application/json:
              schema:
                type: object
                required: [settings]
                properties:
                  settings: { $ref: '#/components/schemas/CollectionSettings' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    patch:
      tags: [Collections]
      operationId: patchCollectionSettings
      summary: Update a collection's settings
      description: |
        Requires `admin` role. Partial update: send only the keys to change. `modules`,
        `theme`, `pluginSettings` and `visibility` shallow-merge; arrays and scalars
        replace. Validation happens before any write, so a rejected request changes
        nothing. Returns the full updated resource.
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CollectionSettings' }
      responses:
        '200':
          description: The updated settings.
          content:
            application/json:
              schema:
                type: object
                required: [settings]
                properties:
                  settings: { $ref: '#/components/schemas/CollectionSettings' }
        '400':
          description: Invalid or empty patch.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { success: false, error: 'Unknown module: vinyl' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `POST /auth/login` (15-minute lifetime).
  parameters:
    CollectionId:
      name: id
      in: path
      required: true
      description: Collection id (Mongo ObjectId).
      schema: { type: string }
...
    ItemId:
      name: itemId
      in: path
      required: true
      description: Item id (Mongo ObjectId).
      schema: { type: string }
...
  responses:
...
    Unauthorized:
      description: Missing, invalid or expired Bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing:
              value: { success: false, error: 'Missing bearer token' }
            invalid:
              value: { success: false, error: 'Invalid or expired token' }
    Forbidden:
      description: Authenticated, but lacking the required role or instance-admin flag.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: { success: false, error: 'Forbidden' }
    NotFound:
      description: The named resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: { success: false, error: 'Collection not found' }
...
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: { success: false, error: 'Internal server error' }
    ServiceUnavailable:
      description: The instance has not completed first-run setup; all `/api/v1` routes return this.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: { success: false, error: 'Instance not set up yet' }
  schemas:
    Error:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error: { type: string }
        code:
          type: string
          description: Optional machine-readable code; not used by every handler.
      example: { success: false, error: 'Invalid email or password' }
...
    ValueSnapshot:
      type: object
      properties:
        capturedAt: { type: string, format: date-time }
        value: { type: number }
        minValue: { type: number }
        maxValue: { type: number }
        itemCount: { type: integer }
    ValueHistory:
      type: object
      required: [currency, snapshots, otherCurrencies]
      properties:
        currency: { type: string, enum: [EUR, USD, GBP] }
        snapshots:
          type: array
          items: { $ref: '#/components/schemas/ValueSnapshot' }
        otherCurrencies:
          type: array
          items: { type: string }
    PriceEstimate:
      type: object
      required: [source, price, details]
      properties:
        source: { type: string, enum: [market, history] }
        price:
          type: object
          required: [value, currency]
          properties:
            value: { type: number }
            currency: { type: string }
        details: { type: string }
    ValueEstimateJob:
      type: object
      required: [id, status, progress]
      properties:
        id: { type: string }
        status: { type: string, enum: [running, done, error] }
        progress:
          type: object
          required: [processed, total]
          properties:
            processed: { type: integer }
            total: { type: integer }
        result:
          type: object
          properties:
            value: { type: number }
            minValue: { type: number }
            maxValue: { type: number }
            itemCount: { type: integer }
            currency: { type: string }
            pricedCount: { type: integer }
            failedCount: { type: integer }
            saved: { type: boolean }
        error: { type: string }
...
    CollectionSettings:
      type: object
      properties:
        modules:
          type: object
          description: One entry per registered plugin's collectionType.
          additionalProperties: { type: boolean }
        mergeDuplicates: { type: boolean }
        pluginSettings:
          type: object
          additionalProperties:
            type: object
            additionalProperties: true
        visibility:
          type: object
          properties:
            applyToAdmin: { type: boolean }
            hiddenItems:
              type: array
              items: { type: string }
            hiddenGenres:
              type: array
              items: { type: string }
            hiddenTypes:
              type: array
              description: Plugin `kind`s (e.g. `Music`), not collectionTypes.
              items: { type: string }
        theme:
          type: object
          description: '`home` or a collectionType, mapped to a preset key.'
          additionalProperties: { type: string }
        aspectRatioClass: { type: string }
        navbarShortcuts:
          type: array
          items: { type: string }
        statsWidgets:
          type: array
          items: { type: string }
        fastAdd: { type: string }
...